pyproject.toml
[tool.poetry]
name = "arango_wal"
version = "0.1.0"
description = "Listen to ArangoDB's WAL changes"
authors = ["Adolfo Silva <code@adolfosilva.org>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/adolfosilva/arango_wal"
repository = "https://github.com/adolfosilva/arango_wal"
keywords = ["arango", "arangodb", "wal", "replication", "listener", "subscriber"]
classifiers = [
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Topic :: Database",
  "Development Status :: 1 - Planning"
]
include = [
  "LICENSE",
]

[tool.poetry.scripts]
example_http_server = "examples.http_server:main"

[tool.poetry.dependencies]
python = "^3.9"
python-arango = "^7.3.4"
pyee = "^9.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
isort = "^5.10.1"
black = "^22.3.0"
pylint = "^2.13.8"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
